rest project doctors

+++ petstagram_softuni_web_basics/pets/views.py
@@ -1,7 +1,9 @@
+from django.contrib.auth.decorators import login_required
 from django.shortcuts import render, redirect
 
 from common.forms import CommentForm
 from common.models import Comment
+from core.clean_up import clean_up_files
 from pets.forms import CreatePetForm
 from pets.models import Pet, Like
 
@@ -13,11 +15,16 @@ def list_pets(req):
     return render(req, 'pets/pet_list.html', context)
 
 
-def get_detail_context(pet):
+def get_detail_context(req, pet):
     comments = pet.comment_set.all()
     return {
         'pet': pet,
         'form': CommentForm(),
+        'can_delete': req.user == pet.user.user,
+        'can_edit': req.user == pet.user.user,
+        'can_like': req.user != pet.user.user,
+        'can_comment': req.user != pet.user.user,
+        'has_liked': pet.like_set.filter(user_id=req.user.userprofile.id).exists(),
     }
 
 
@@ -26,13 +33,14 @@ def show_pet_detail(req, pk):
     pet.likes_count = pet.like_set.count()
 
     if req.method == 'GET':
-        return render(req, 'pets/pet_detail.html', get_detail_context(pet))
+        return render(req, 'pets/pet_detail.html', get_detail_context(req, pet))
 
     elif req.method == 'POST':
         form = CommentForm(req.POST)
         if form.is_valid():
             comment = Comment(comment=form.cleaned_data['comment'])
             comment.pet = pet
+            comment.user = req.user.userprofile  # do not link to profile but user
             comment.save()
             pet.comment_set.add(comment)
             pet.save()
@@ -43,14 +51,20 @@ def show_pet_detail(req, pk):
         return redirect('pet_details', pet.id)
 
 
+@login_required
 def like_pet(req, pk):
-    pet = Pet.objects.get(pk=pk)
-    new_like = pet.like_set.create()
-    # like = Like(pet=pet)
-    # like.save()
-    # pet.like_set.add(like)
-    # pet.save()
-    return redirect('pet_details', pk)
+    #checks if liked and dislikes
+    like = Like.objects.filter(user_id=req.user.userprofile.id, pet_id=pk).first()
+    if like:
+        like.delete()
+    else:
+        pet = Pet.objects.get(pk=pk)
+        # new_like = pet.like_set.create()
+        like = Like(pet=pet, user=req.user.userprofile)
+        like.save()
+        pet.like_set.add(like)
+        pet.save()
+        return redirect('pet_details', pk)
 
 
 def persist(req, pet, html_template, redirect_url):
@@ -61,7 +75,10 @@ def persist(req, pet, html_template, redirect_url):
         return render(req, f'pets/{html_template}.html', context)
 
     elif req.method == 'POST':
-        form = CreatePetForm(req.POST, instance=pet)
+        old_image = pet.image
+        if old_image:
+            clean_up_files(old_image.path)
+        form = CreatePetForm(req.POST, req.FILES, instance=pet)
 
         if not form.is_valid():
             context = {
@@ -76,23 +93,29 @@ def persist(req, pet, html_template, redirect_url):
         return redirect(redirect_url)
 
 
+@login_required
 def create_pet(req):
     return persist(req, Pet(), 'pet_create', 'list_pets')
 
-
+@login_required
 def edit_pet(req, pk):
     pet = Pet.objects.get(pk=pk)
     return persist(req, pet, 'pet_edit', 'pet_details')
 
 
+@login_required
+#@require_user(model=Pet)
 def delete_pet(req, pk):
     pet = Pet.objects.get(pk=pk)
-    if req.method == 'GET':
-        return render(req, 'pets/pet_delete.html', {
-            'name': pet.name
-        })
+    if pet.user.user != req.user:
+        pass
+    else:
+        if req.method == 'GET':
+            return render(req, 'pets/pet_delete.html', {
+                'name': pet.name
+            })
 
-    elif req.method == 'POST':
-        pet.delete()
-        return redirect('list_pets')
+        elif req.method == 'POST':
+            pet.delete()
+            return redirect('list_pets')
